pyproject.toml
[project]
name = "vaultar"
version = "0.1.0"
description = "Secure backup and restore CLI tool"
authors = [{ name = "Marcelo Trindade", email = "marcelost@riseup.net" }]
dependencies = [
    "click>=8.0.0",
    "rich>=12.0.0",
    "python-gnupg>=0.5.0",
    "zstandard>=0.19.0",
]
requires-python = ">=3.8"

[project.scripts]
vaultar = "vaultar.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
